+++ src/pages/ContentScripts/index.ts
@@ -3,6 +3,7 @@ import { initializeIcons } from '@fluentui/react/lib/Icons';
 import DeveloperActiInflTrend from '../../anchors/DeveloperActInflTrend';
 import RepoActiInflTrend from '../../anchors/RepoActiInflTrend';
 import PerceptorTab from '../../anchors/PerceptorTab';
+import PerceptorLayout from '../../anchors/PerceptorLayout';
 import DeveloperNetwork from '../../anchors/DeveloperNetwork';
 import { inject2Perceptor } from './Perceptor';
 initializeIcons();
@@ -10,7 +11,7 @@ inject2Perceptor(DeveloperActiInflTrend);
 inject2Perceptor(RepoActiInflTrend);
 
 inject2Perceptor(PerceptorTab);
-import './PerceptorLayout';
+inject2Perceptor(PerceptorLayout);
 inject2Perceptor(DeveloperNetwork);
 import './ProjectNetwork';
 import './Hypertrons';
